Law_pipeline_reddit_news_v2/requirements.txt
scrapy==2.8.0
playwright>=1.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
praw>=7.7.1
asyncpraw>=7.7.1
beautifulsoup4==4.12.2
requests==2.31.0
pytest>=7.4.0
pymongo>=4.5.0
airbyte>=0.5.0
selenium>=4.15.0
webdriver-manager>=4.0.1
sqlalchemy>=2.0.0
#psycopg2-binary==2.9.6
psycopg2-binary>=2.9.9,<3.0.0
sqlalchemy-utils>=0.41.1
trafilatura==1.6.0
configparser==5.3.0
pathlib==1.0.1 